% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{backwards}
\alias{backwards}
\title{Shift start time of potential mortalities earlier}
\usage{
backwards(data, morts, ID, station, res.start, season = NULL, stnchange = NULL)
}
\arguments{
\item{data}{a dataframe of residence events. Residence events must include
tag ID, location name, start time, and duration. Residence events should be
continuous (i.e., not subset by season).}

\item{morts}{a dataframe containing potential mortalities. Must use
#' the same column names and in the same order as \code{data}.}

\item{ID}{a string of the name of the column in \code{data} that holds the tag or
sample IDs.}

\item{station}{a string of the name of the column in \code{data} that holds the
station name or receiver location.}

\item{res.start}{a string of the name of the column in \code{data} that holds the
start date and time. Must be specified and in POSIXt if \code{format="manual"}.}

\item{season}{a dataframe with the start and end times of the season(s) or
time period(s) of interest.}

\item{stnchange}{a dataframe with the start time and location of the most
recent station or location change. Must use the same column names and in
the same order as \code{data}.}
}
\value{
a dataframe with one row for each tag ID, including the date/time of
the residence start when the potential mortality or expelled tag was identified.
Returns the input \code{morts} dataframe if no potential mortalities are shifted
earlier.
All input data fields
(e.g., any name, location, or species information that was included with the
input data) will be retained.
}
\description{
Shift the start time of potential mortalities earlier, if
station/location has not changed.
}
\examples{
\dontrun{backwards(data=data,morts=morts,ID="TagID",station="Receiver",
res.start="ResidencesStart",stnchange=stn.change)}
}
