% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddd}
\alias{ddd}
\title{Sample dead drift direction dataframe}
\format{
A data frame with 18 rows and 2 variables.
}
\usage{
data(ddd)
}
\description{
The locations have been given generic station names, due to the sensitive nature
of the dataset. The connections between stations (i.e., which stations are
connected by drift and in which direction) are real.
}
\details{
The variables are as follows:

\itemize{
\item From. The station where a tag may drift from.
\item To. The station where a tag may drift to, from the station in From.
}
}
\keyword{datasets}
