% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detections}
\alias{detections}
\title{Sample acoustic telemetry detection data}
\format{
A data frame with 447 627 rows (observations) and 3 variables.
}
\usage{
data(detections)
}
\description{
These detection data were subset from a real acoustic telemetry dataset of Arctic
Char near Kugluktuk, Nunavut. The data were from a project that was a
collaboration between the Kugluktuk Hunters and Trappers Organization, the
University of Waterloo, and Fisheries and Oceans Canada. The data are valuable
to the community of Kugluktuk, and are also sensitive due to their pertinence
to the local fishery for Arctic Char. For this reason, the detections have been
given generic station names and fish IDs so locations and sample information
is anonymous. The year has also been changed, but otherwise, the detections are
real, and therefore have all the challenges and intricacies of biological
data.
}
\details{
The variables are as follows:

\itemize{
\item DateTimeUTC. The date and time of each detection, in POSIXct format. The
data are in UTC, as downloaded from the receivers.
\item Station.Name Name of the receiver location.
\item ID Unique ID of the fish.
}
}
\keyword{datasets}
