% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{format.auto}
\alias{format.auto}
\title{Fill in field names automatically}
\usage{
\method{format}{auto}(format, field, data)
}
\arguments{
\item{format}{the format of the dataset to be analyzed. Options are "mort",
"actel", "glatos", and "vtrack" (note lowercase "v").}

\item{field}{the argument for which a field name will be automatically
generated. Options are "res.start", "res.end", "residences", and "units".}

\item{data}{}
}
\value{
a character string with the appropriate name of the field, to be used
as an argument in further analyses in \code{mort}.
}
\description{
Fills in field names automatically for dataframes of residence
events that were generated using common packages for processing acoustic
telemetry data. Supported packages are \code{mort}, \code{actel}, \code{glatos}, and
\code{Vtrack}.
}
\examples{
format.auto("actel","units")
format.auto("glatos","res.start")
format.auto("mort","res.end")
format.auto("vtrack","residences")
}
