% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mort.R
\name{mort}
\alias{mort}
\title{MortTemp}
\usage{
mort(data, type = "mort", units, residences)
}
\arguments{
\item{data}{a dataframe with residence events.}

\item{type}{the format used to generate the residence events. Options are
"mort", "actel", "glatos", "vtrack", or "manual". If "manual", then user
must specify the \code{units} and \code{residences}.}

\item{units}{units of the duration of the residence events in \code{data}. Must be
specified if \code{type="manual"}.}

\item{residences}{a character string with the name of the column in \code{data}
that holds the duration of the residence events. Must be specified if \code{type=manual}.}
}
\value{
a dataframe with...
}
\description{
MortTemp
}
\examples{
\dontrun{mort(data=res.events,type="manual",units="days",residences="ResidenceLength")}
}
