% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residences.R
\name{residences}
\alias{residences}
\title{Generate residences}
\usage{
residences(data, ID, cutoff, units)
}
\arguments{
\item{data}{a data frame of detection data.}

\item{ID}{a string of the column name in data frame \code{data} that includes the tag or sample ID}

\item{cutoff}{the maximum allowable time difference between detections to
be considered a single residence event.}

\item{units}{the units of the cutoff. Options are "secs", "mins", "hrs", or "days".}
}
\value{
A data frame
}
\description{
Generate residences
}
\examples{
residences(data=dat,cutoff=1,units="days")
}
