% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seasons}
\alias{seasons}
\title{Ice-free seasons}
\format{
A data frame with 5 rows (years) and 2 variables.
}
\usage{
data(seasons)
}
\description{
The year has been changed, due to the sensitive nature of the dataset. The
years have been changed in the same manner as the detection dates, so
the ice-free dates correspond to the dataset in the same way as the original
dataset.
}
\details{
The variables are as follows:

\itemize{
\item Start. The date of ice break-up (the start of the period of interest).
\item End. The date of ice freeze-up (the end of the period of interest).
}
}
\keyword{datasets}
