% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mort.R
\name{morts}
\alias{morts}
\title{Identify potential mortalities or expelled tags}
\usage{
morts(
  data,
  format = "mort",
  ID,
  station,
  res.start = "auto",
  res.end = "auto",
  method = "all",
  units = "auto",
  residences = "auto",
  season = NULL,
  singles = FALSE
)
}
\arguments{
\item{data}{a dataframe of residence events. Residence events must include
tag ID, location name, start time, and duration. Residence events must also
include end time if \code{season} is provided.}

\item{format}{the format used to generate the residence events. Options are
"mort", "actel", "glatos", "vtrack", or "manual". If "manual", then user
must specify \code{ID}, \code{station}, \code{res.start}, \code{res.end}, \code{residences}, and \code{units}.}

\item{ID}{a string of the name of the column in \code{data} that holds the tag or
sample IDs.}

\item{station}{a string of the name of the column in \code{data} that holds the
station name or receiver location.}

\item{res.start}{a string of the name of the column in \code{data} that holds the
start date and time. Must be specified and in POSIXt or character in the format
YYYY-mm-dd HH:MM:SS if \code{format="manual"}.}

\item{res.end}{a string of the name of the column in \code{data} that holds the
end date and time. Must be specified and in POSIXt or character in the format
YYYY-mm-dd HH:MM:SS if \code{format="manual"}.}

\item{method}{the method to be used in flagging mortalities. Options are
"last", "any", "cumulative", or "all"}

\item{units}{units of the duration of the residence events in \code{data}.}

\item{residences}{a character string with the name of the column in \code{data}
that holds the duration of the residence events.}

\item{season}{a dataframe with start and end dates of the season(s) of interest}

\item{singles}{specifies if single detections (length of residence event = 0)
should be removed. Removing single detections is the most conservative method,
so chance or potentially invalid detections do not affect mortality estimates.}
}
\value{
a dataframe with one row for each tag ID, including the date/time of
the residence start when the potential mortality or expelled tag was identified.
All input data fields
(e.g., any name, location, or species information that was included with the
input data) will be retained.
}
\description{
Identifies potential mortalities or expelled tags from passive
acoustic telemetry data. Mortalities are identified based on thresholds
derived from the dataset itself.
}
\examples{
\dontrun{mort(data=res.events,format="manual",units="days",residences="ResidenceLength")}
}
